Added pagination to Table
Added pagination to the table component. Correctly displays data on the bottom left and buttons as well as previous and next buttons on the bottom right. Uses my default button style so I may change that in the future to something smaller. Has a small bug that I'll need to fix with useEffect when you're on a page and then search but the results have less pages than you were currently on.